+++ apps/docs/app/examples/table/use-case/page.tsx
@@ -544,7 +544,7 @@ export default function Page() {
             )}
           </TableBody>
         </Table.Content>
-        <Table.Footer>{bottomContent}</Table.Footer>
+        <div>{bottomContent}</div>
       </Table>
     </div>
   );