+++ apps/docs/content/components/table/custom-styles.raw.tsx
@@ -659,7 +659,7 @@ export default function App() {
           )}
         </TableBody>
       </Table.Content>
-      <Table.Footer>{bottomContent}</Table.Footer>
+      <div>{bottomContent}</div>
     </Table>
   );
 }