+++ apps/docs/content/components/table/load-more.raw.jsx
@@ -74,7 +74,7 @@ export default function App() {
           )}
         </TableBody>
       </Table.Content>
-      <Table.Footer>
+      <div>
         {hasMore && !isLoading ? (
           <div className="flex w-full justify-center">
             <Button isDisabled={list.isLoading} variant="flat" onPress={list.loadMore}>
@@ -83,7 +83,7 @@ export default function App() {
             </Button>
           </div>
         ) : null}
-      </Table.Footer>
+      </div>
     </Table>
   );
 }