+++ apps/docs/content/components/table/use-case.raw.jsx
@@ -634,7 +634,7 @@ export default function App() {
           )}
         </TableBody>
       </Table.Content>
-      <Table.Footer>{bottomContent}</Table.Footer>
+      <div>{bottomContent}</div>
     </Table>
   );
 }